+++ lib/sbmt/kafka_consumer/routing/karafka_v2_consumer_mapper.rb
@@ -1,0 +1,9 @@
+module Sbmt
+  module KafkaConsumer
+    module Routing
+      # uses default karafka v2 mapper
+      # exists just for naming consistency with KarafkaV1ConsumerMapper
+      class KarafkaV2ConsumerMapper < SbmtKarafka::Routing::ConsumerMapper; end
+    end
+  end
+end